Initial Release
This commit contains the first working release for Daily-Vaults-MC. In retrospect, I should have created this repository earlier in the project, but I didn't even know if this was going to work. Don't worry, though, as this is just the initial version, and more updates will soon follow (including a logo, README, and compatibility with other Minecraft: Java versions).